# [Mono.Android] Use the faster java type name mapping