This fixes https://github.com/xamarin/xamarin-android/issues/1831

Make sure we use the code path, which uses the `typename.*` maps. It
is much faster than `JavaNativeTypeManager.ToJniName`.

Tested with x86 emulator again. It saves 17ms (from 18ms to 1ms) in XA
template and 26ms (from 28ms to 2ms) in XamlSamples app. So it scales
well.

It also decreases JIT usage. XA template: *Compiled methods* 1954 to
1880.